refactor update password and change password forms
